Add files via upload
The css files need to be kept together in a folder named css and the java file in a file named js . The main page is the cookbook page and the cookie, donuts and panna cotta file is the sub interactive pages. You will find the dish recipes linked to the images. Thank you. ; )